+++ packages/desktop/scripts/download_pyodide.sh
@@ -1,0 +1,6 @@
+#!/bin/bash -eu
+
+PYODIDE_VERSION=$(node -e 'const { version } = require("pyodide"); console.log(version)')
+echo "Downloading Pyodide $PYODIDE_VERSION"
+mkdir -p ./build
+curl -fSL https://github.com/pyodide/pyodide/releases/download/$PYODIDE_VERSION/pyodide-core-$PYODIDE_VERSION.tar.bz2 | tar xj -C ./build --files-from=./pyodide-files.txt