+++ packages/sharing-editor/public/samples/011_component_gallery/pages/widget.multiselect.py
@@ -3,7 +3,7 @@ import streamlit as st
 options = st.multiselect(
     "What are your favorite colors",
     ["Green", "Yellow", "Red", "Blue"],
-    ["Yellow", "Red"],
+    default=["Yellow", "Red"],
 )
 
 st.write("You selected:", options)