+++ packages/sharing-editor/public/samples/011_component_gallery/pages/widget.multiselect_accept_new_options.py
@@ -1,0 +1,10 @@
+import streamlit as st
+
+options = st.multiselect(
+    "What are your favorite cat names?",
+    ["Jellybeans", "Fish Biscuit", "Madam President"],
+    max_selections=5,
+    accept_new_options=True,
+)
+
+st.write("You selected:", options)